Fix: Ubuntu Google web font styles
Previous commit included changes to main.css that is not included in the
head.html page and therefore the style definitions are not used.

Added slight increase to pre font size to help code stand out.

Added bigger margin at bottom of pre sections.

Added styling for blockquote to make the hint sections look different
to the rest of the content.